+++ package/Changelog.md
@@ -4,5 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-pre] - 2022-08-07
+- fixed class and namespace names
+- added package author
+- added timeline binding helpers
+
 ## [0.0.1-pre] - 2022-08-07
 - Initial release